HYD-2008: Fix test_saved_outlet_triggers_fencing_update
The previous test was unreliable and racy. This version
removes the possibility of races and more directly tests
the effects of outlet <=> host association by looking
at the output of fence_chroma -o list on the test server.

This commit also adds new agent actions to put a Pacemaker node
into and out of administrative standby to better signal our
intentions to Pacemaker and improve its predictability. This
functionality will be exposed via the API with HYD-2071.

Change-Id: I33a66b80e21909f201a6c58412604117101ee1a9