HYD-2116: Cope with updates that have new centos dependencies
This modifies the behaviour of update_packages such that it queries
the specified repos for updates, and then executes the update operation
with all repos enabled.  The result is that only new packages in the
desired repos (IML repos) trigger updates, but if those updates
have dependencies from other repos then they are installed
successfully too.

Change-Id: If44b70dd456182cf5c84776bbeb331894dadc963