HYD-2338: Failure to start targets after LNet unload
During 2.0 RC testing, it was discovered that the following
sequence would cause targets on the affected server to fail
to restart:

Stop LNet
Unload LNet
Load LNet
Start LNet
Attempt to start a target on the server which had unloaded LNet

The problem seems to be that 49e8308 reverted a workaround that had
previously landed in ba5e8d0. This commit just reinstates the
workaround.

Change-Id: I8a45dd400f1fc9f84e66bb9a803d2fdff459e4a2