HYD-2578 md-raid devices are not supported by IML
md-raid devices caused an exception to be thrown in the agent.

File "/usr/lib/python2.6/site-packages/chroma_agent/device_plugins/linux.py", line 251, in _init_
KeyError: '/dev/md127'

This the problem was that device path was that node_block_devices was
being access with a path to fetch a major_minor but actually the major_minor
was already present - and node_block_devices is index by a major_minor not
a path.

Most of the work - 90% was to create some tests to stop this regressing,
as well as adding some resilience to the code so that exceptions and bad
data are not thrown/returned if things go bad.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: I8a4f8498ef82a7f11e2c4ce4ac49920c995ef8af