HYD-2771 Query configured network and lnet interfaces
Query network and lnet interfaces, allow update via a rest interface.

The code provides a rest interface to allow networks to be viewed
and nids to be associated with those networks.

The rest interface allows for nids to be delete using delete or
deleted by posting with a network_no of -1. This means that complete
updates for a host can be performed atomically. This is definately
the prefered method.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: I89ae71191a2901bacfb575bdd26ddac3124d9729