HYD-2863 Test failure: integration test, No JSON object could be decoded
This patch adds additional debug around the failing line of code, hopefully
this will produce the required debug to resolve the issue and if not
additional debug to move use in the right direction.

Also contains a change so the client logs are not deleted after each test.
This hopefully makes the client logs useful.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: Ibf601179bf47098c29a2d6185655fb9ccbccf4ca