HYD-2984: Copytools should use correct mountpoint
The code that configures copytools was using an older method
which resulted in the copytool looking in the wrong place
for a mounted lustre filesystem. This commit corrects that
problem and ensures that the copytool is looking at the
mountpoint supplied by the user during copytool creation.

Also adds validations for the copytool arguments to flag known
problems (e.g. --archive is redundant, --daemon is wrong, etc.).

Change-Id: I9ffadbc53607f75662f99fadc59776b4bcb4f9ae