HYD-2950 Firewall rule for ntp not re-created after removing and re-installing chroma manager
This change regularises and removes a special case where the
ntp server to be configured is the same as the configured server.
By removing a special case 'return' and making the path the same
for all cases the issue of the firewall rule not being created
it removed.

I also took the opportunity to tidy some naming of the files into
variables so that we removed some magic strings from the code.

This is related to ntp and so I think a fair small improvement
for this patch.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: I19d65eae7e1ae183df40f1dcd9d18edda09514b3