HYD-3055 Testing on Regal demonstrated an issue with o2ib and the new LNet
The issue turned out to be that the o2ib code is not tested properly meaning
the code had o2ib regex code issues in it. It was also apparent that the
code did not cause an lnet alert when the nids changed. A modification
was added to resolve this.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: I2dc60b99f399040c3247705b9932ef970da46a7f