HYD-2916 Fix client count
lustre changed the num_exports value in the MDT to mean the number of total
connections to the MDT, which would include 2 internal connectiosn and
the total number of OST connections + Client connections.  To find the
number of connections we must calculate the number of OST connectiosn
That's done via:
  lctl get_param lov.$fs_name-*.target_obd
The output must then be parsed to count the number of OSTs
The number of exports is then adjusted to count clients only:
  num_exports = num_exports - (2 + OST count)

The unit-tests have also been updated to account for the change.
Existing unit tests already test the new code.

Change-Id: Ib19ab90b149c3212e3613eaa0c7cce3aa4824f13
Signed-off-by: Amir Shehata <amir.shehata@intel.com>