HYD-2916 Fix client count
original solution: use lctl get_param to get the number of OSTs
and then calculate the number of clients, based on num_exports
CONS: lctl get_param considers OSTs ACTIVE even if they are not
active (or trying to connect to them - IE unreachable).
However, num_exports reflects the correct number of active
connections.  Thus the client count calculation gets messed up.

solution in this patch: iterate through the mdt/<fs>-MDTXXXX/exports/
for all the directories (1 per nid connected + local nid), check if
there is a uuid file, and ignore all entires with "<fs>-MDT" substring.
All other entries are for legitimate clients.

CONS:  LU-4596.  The uuid file is not refrehsed if the client
un mounts and re mounts.  this will lead to client count being
in accurate.  However LU-4596 has been fixed and will be rolled in
the IEEL lustre.

the good thing about proposed solution is it can be modified easly to
not only show the number of clients currently mouting the file system,
but it can also list their nids, if that's a desired feature.

added a test case for the newly added function: get_client_count()

Change-Id: I66060d23ff83dafef3c5e1f403f52714e4974fbb
Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Reviewed-on: http://review.whamcloud.com/9895
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Aric Coady <aric.coady@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>