HYD-3093: Don't fail immediately if cib is unavailable
In HYD-3113, a change was landed to make the PacemakerConfig
class check for CIB availability on instantiation. The goal was
to fail early when pacemaker isn't running. The problem is that
while this works as intended most of the time, there is now a
race after restarting pacemaker.

Instead of failing immediately, we should retry, but for a much
shorter period of time than usual. 10 seconds seems reasonable
as a timeout for waiting for pacemaker to be available after a
restart.

Change-Id: I63ae4b3a060bbc014f1e4803e66e1667833c0895
Reviewed-on: http://review.whamcloud.com/10019
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>