HYD-3139 Host add hung - UpdateDevices step in Load LNet module
The problem is that the normalized devices routine would allow circular
references to be entered, i.e. A normalized to A this then causes a hang
when we try and retrieve them.

As an additional measure I added a counter to detect larger circular
references such as A->B->C->A in this case we don't know which is the
normalized value so just drop out once we see that we are about to
got back to the starting value.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: I7dd8489d0af4a04ceeaf8a91b1b41c92d649db1e
Reviewed-on: http://review.whamcloud.com/10065
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>