HYD-3172 Device detection throws errors due to unnecessary normalization
The lustre device plugin normalizes devices so that the major:minor
table doesn't actually relate to the actual device.

The fact is that because we build a complete stack and normalize
the devices durring use it is simpler and cleaner to keep the
relationship between major:minor a true representation of the
device.

Whilst this seems to undo HYD-1090 I actually believe HYD-1090
can't be solving the problem because it only normalizes for
standard linux devices such a multi-mapper, it can't be
achieving what it wanted for things like mdraid, lvm, emcpower
etc. As those devices work the normalize worked cannot be required

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: I171090af270eb67318474417f0afd42ef2f82bad
Reviewed-on: http://review.whamcloud.com/10173
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>