HYD-3172 Device detection throws errors due to unnecessary normalization
The lustre device plugin normalizes devices so that the major:minor
table doesn't relate to the actual device.

The fact is that because we build a complete stack and normalize
the devices durring use it is simpler and cleaner to keep the
relationship between major:minor a true representation of the
device.

However I don't want to undo HYD-1090 just before an RC so I have a
half way house solution that still uses the old function for this single
use in linux.py

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: Ia90abeafb761b9369795525250c0157b7880d850
Reviewed-on: http://review.whamcloud.com/10182
Tested-by: Jenkins
Reviewed-by: Aric Coady <aric.coady@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>