HYD-2929 Detect ZFS filesystem on monitored system deal with appropriately.
Until managed ZFS lands we need to detect in a monitored filesystem that
zfs is in use and tell the user the appropriate message.

Given to do this we need to detect ZFS, this ticket amounts to all of ZFS
support other than the format. Which means we could land the ZFS ticket
with the ability to format ZFS removed.

To do this we would need to flag to the GUI that a device is detected
and visible but not selectable for use in a new files system.

Change-Id: I0f69d6abfe8047bdc7de8d93fb44b33cdea784a3
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/11179
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Reviewed-by: Aric Coady <aric.coady@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>