HYD-2614 Implement changes that enable ZFS and others to be added (Patch 2)
The changes to generalise IML's code to allow different backend filesystems
should be carried out under this ticket. This ticket will not include any
zfs additions and so functionally when this ticket is completed IML will
be the same.

However adding zfs should then be a case of adding ZFS code
and classes and not changing core IML code.

Key elements are;

1. Formatting and mounting a filesystem
2. Device detection
3. Existing Filesystem detection
4. HA Failover

This change deals with 3.

Change-Id: If63a6b6d680e73d771e9cfd03fa85773221b02fc
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/11192
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Reviewed-by: Aric Coady <aric.coady@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>