HYD-2929 Detect ZFS filesystem on monitored system
This patch fixes up some issues with filesystem detection
on zfs. Fundamentally it was broken and my local testing
only passed because I created the filesystem with IML and
then deleted and redetected it. This meant the debugfs code
which spots the mgs didn't need to work.

This code moves the blockdevice specific code into the appropriate
classes and then uses instances of the classes to work.

There are another couple of fixups for audit and stats that
integration testing spotted. I also added some tests for the
BlockDeviceZFS class as it is now doing a little computation.

Change-Id: I401a47bb273602c3184a02a98fb56298f456d75d
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/11917
Tested-by: Jenkins
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>