HYD-3629 Further enhancements to ZFS HA Monitoring
HYD-3629 required some changes the caused difficulty with
HA detection. This patch fixes up those changes and actually
makes the zfs HA detection code much simpler.

The strategy is that exported ZFS pools are imported read only
and then treated as if local before being exported. Whilst
this operation is relatively slow it only happens occasionaly
such as during detection and we would not be expecting any
management operations to be occuring on exported pools.

Change-Id: If27cd2b81f931fb12b1a11d8a7dd7f5b61c50d52
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/12168
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>