HYD-3727 Targets incorrectly alerted to be offline for lvm
lvm devices are incorrectly reported offline by iml. This is because
the UUID returned by the device detection is lvm-volume UUID but the
UUID returned by the lustre audit comes from Blkid and this are not
the same

The solution is to create a lvm_volume device so that the BlkId read
for lvm volumes can correctly read the lvm volume uuid.

Change-Id: Ia6bf65b49a5618f9957456c1d0b8bce83fc06e17
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/12348
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>