HYD-3576 Stats tests are failing with ZFS
Discussions with Johann indicate that all the values gives by Lustre
are reasonable estimates that the filesopen will be accurate. However
freespace and total space will vary are space is used up for inodes
and inodes free is a good estimate.

This patch ensures IML reports these reasonable values by allow for tolerance
in the case of zfs.

Change-Id: Ib90e118186c4ee251e4b2478e15734d128d61086
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/12569
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>