HYD-3855 Intermittent failure setting host profile, missing iml-agent file in yum cache
This patch which may be part of a number of patches fixes up the 'TypeError: __init__()
takes at least 5 arguments (2 given)' part of this bug. It doesn't address the yum error
but makes it prettier.

As a part of this fixup I added some tests and also re-organized the install process so
that error detection in the profile setup causes the reset of the install to gracefully
stop.

Finally a simple refactor in JobSchedular pulls out a lot of repetition, this isn't
required and causes no functional change. I made the change during experimentation
and it is a step forward, DRY and all that, so it seemed odd to pull it out again.

Change-Id: Ib660eb0e59308964ac547805114f9c4db71428d7
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/12930
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Michael MacDonald <michael.macdonald@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>