HYD-3696: Heatmap not showing new OSTs
The root cause of this issue (along with many others) is that there are
no read/write stats entries created for new targets until there has been
some I/O activity involving the target.

This patch adds some logic in the agent-side audit to provide default
values for specific stats. For the moment, the only stats we want to
provide defaults for are read_bytes and write_bytes.

Change-Id: Ica4f08b2297f68da9a7486d6a26d3f966934a046
Reviewed-on: http://review.whamcloud.com/12989
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>