HYD-3909 ZFS presence tests blocks on stdin and hangs chroma-agent
When running chroma-agent from the command line with

the which zfs test would fail because it had -i which reads aliases
from stdin. As stdin was not provided the app would hang.

The -i is not a requirement and so simply removing -i resolves this
issue.

As well as this change the unit test is changed to mock out the which zfs
command rather than test for the presence of zfs on the test machine.

Change-Id: I737aa6b07cbfe35ca3aa60e34019d042a2bd6706
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13037
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>