HYD-3905 TestActionRunnerPlugin, test_two_actions diagnosis
A race condition in the mock code caused the value of call count
to not be correctly incremented and occasionally loose a value.

This patch changes from the generic mock which is not threadsafe
to a specific mock which is thread safe.

The test of 2 concurrent actions is changed to a test of 1000 concurrent
actions. If 1000 are run on the old code then the failure is seen everytime.

There are no changes to the product code, this patch fixes a test bug.

Change-Id: Iaecc60f13c3965e53546914d12d6b9c687308ef0
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13266
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>