HYD-4040 Add support for fence_apc_snmp
This adds support for the APC AP8941 which is an
SNMP based PDU. The change is untested as we don't
have the hardware to test with.

We are adding only this particular model because
that is the limit of the support from RedHat.

Being snmp based this model uses the fence_apc_snmp
agent. and it is because of this change of agent that
a code rather than data change is required.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I252879f02ad184375b71510c86274e982e69e133
Reviewed-on: http://review.whamcloud.com/13471
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>