HYD-4048 Testing improvements to shell patching
On the agent side (and the manager) there are 3 ways of patching the
shell commands. We should be aiming for 1 way in a common library so
we can get the greatest leverage from our work.

Change almost everything to use the CommandCaptureTestCase class to
patch shell commands. This provides for better testing of results. The
class can be extended in the future as we see fit.

Almost everything because this deals with all the cases that used a
library function and one or two others that used unique methods. Some
others still exist and this can be fixed up in the future.

Change-Id: I217636f63533edbcc7b087925793c866a0908907
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13449
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>