HYD-4077 ReadServerURI does not account for Authentication
The agent makes no use of the /api manager endpoint other than
in the ReadServerURI call which is used to fetch the agent profile
from the manager. This is an anonymous call to the api which will
fail if ALLOW_ANONYMOUS_READ is set to False. Fixing this up would
require the agent to login to the manager as a user which is
really impractical. It is also the case that the use of this
quick fix to fetch the profile rather than use of the agent messaging
protocol is something that should be removed.

So the change is that instead of the agent pulling the profile from
the manager the manager pushes the profile to the agent using
the standard agent <-> manager interface.

Change-Id: I60228dd9e8170c418d1e8a4b86b4fe33997ff691
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13547
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>