HYD-4050 Warn the user if new kernel won't boot
If the new kernel installed as part of the install package
step won't boot either because there is no default kernel
in grub or because the initramfs didn't get created, then
warn the user and fail the host add command.

The install_packages agent rpc has been extended to return
either an error or the original scan packages data. The
step can the respond appropriately and either create an
AgentException with the error messaage rather than a backtrace.

The error message is also added to the log and so appear one
the command dialog this means that the user does not need to
click through to the failing step.

unit tests added to check the behaviour of the install_packages.

Change-Id: Idcd9784289ae420948164d7b87ab8f62d7532e89
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13553
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>