HYD-3988 Workaround segfault in python in EL 6.6
The patch to make 'at' a requirement for the management
part of the agent didn't ensure the 'atd' daemon would
be started after the install. The daemon is required for
the 'at' command to work. The 'at' command fails silently
with an rc=0 if the daemon is not running!

Simple at 2 lines to the spec file for the management install
part of the agent to ensure 'atd' step up.

chkconfig atd on
service atd start

Change-Id: If58b320e3db448078899ac701e47bf9941b350d4
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13609
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>