HYD-4491 Make filesystem detection more resiliant for zfs filesystems
ZFS filesystem detection is fragile because users must remember to do
partprobe after carrying out any changes to the zfs pools on paired
HA nodes. For a real system where the HA is up and tested before the
nodes are added to the IML system this would not be an issue but for
test systems where HA is simuated and muct of the work is carried out
by hand it is often the case that changes made are not visible to all
the nodes in the system. This fragility needs to be fixed by probing
partitions before look at zfs attributes and ensuring that all zfs
attributes are updated before filesystems are detected.

The patch does three things - two of which are connected to zfs filesystems
resiliance and one of which is an optimization to minimize hits to the
database. Firstly it rescans all the devices on the targets
before calling detect, this removes any race conditions. Second it does
a partprobe before doing a fullscan of the ZFS devices. Thirdly it
caches (in the true sense of the word)  the value of hosts in the
HostListMixin so that the hosts property hits the database the minimum
number of times.

All these are safe things to do with very low risk, they make the
detect take longer (but it is a one time operation) but improve the results.

  - Cache hosts value be looking for changes in the host_ids variable
  - Add partprobe before the zfs detection routines
  - Scan devices before doing a filesystem scan

Change-Id: Iaec2f3a65a2b6b154bdfabdda671956c6e53ebce
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/15211
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Tested-by: Jenkins
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
(cherry picked from commit 6ba1168d29f30616bb0dec24184e2285301b716e)
Reviewed-on: http://review.whamcloud.com/15338