HYD-4247 SetHostProfileStep: could not contact server
After deployment it can take a long time for the agent
to contact the server. The code currently has a 30 second
wait for the contact which it seems in a test environment
is not long enough. A user would have to be pretty quick on
the feet to see the end of deployment, click next, ignore
the redboxes that are present until the agent connects and
then please install. However a test environment can do this
pretty quick, and doesn't wait for the red boxes.

The fix is to explicitly wait for the agent to connect back
to the manager durring the deploy step and hence making
connection back to the manager a requirement for successful
deploy.

 - removed a bit of debug that has crept into agent_client.py
 - improved debug info and error messages for this issue
 - added and used an await_connection function for use in deploy

Change-Id: If0514b3fb53ed2e5789ec1409f1a8cdbef995f9d
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/16085
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>