HYD-4351 Add functionality to to create ZFS FS in IML managed mode
IML doesn't currently build lustre file systems on zpools in managed
mode.

In order to handle zfs devices within IML, zpools are modelled as
either BlockDevice or FileSystem objects and zfs datasets as
FileSystem objects.

This follows the assumptions that the IML BlockDevice is a device
that can be formatted and the resulting device of the format can be
written to in userspace.

As a result, changes are required to reconcile these mappings between
BlockDevice objects to zpools and FileSystem objects to datasets.

With ZFS, the IML model of "formatting" involves creating a dataset
on top of the parent zpool. Both zpool and dataset can be
represented as a BlockDevice but only the dataset can be represented
as a FileSystem. This is understandable given zpool would not be
written to with user data in normal circumstances (not a FileSystem)
and the dataset can contain nested datasets containing user data
(could be a BlockDevice).

When ZFS formatting occurs within IML we need to update the IML view
of the target to reference the dataset rather than the zpool as both
can be thought of as block devices. After this update (in
UpdateManagedTargetMount), the target object attributes reference
the dataset (e.g. pool1/MGS) rather than the zpool (e.g. pool1) as
expected.

UpdateManagedTargetMount is a new step of FormatTargetJob needed
to update mtm volume_nodes for a target so that datasets added during
MkfsStep are referenced rather than the zpool they were created on.

Inherit BlockDevice subclasses from their peer FileSystem subclasses
to enable the use of inherited methods. It is envisaged that
FileSystem and BlockDevice classes will be merged in future given the
questionable benefit of the different abstractions.

Due to multiple inheritance, changed FileSystem.__new__ to create new
class instance with call to object.__new__ rather than
Super(FileSystem, cls).__new__ to avoid BlockDevice.__new__ being
called (as its lower in the MRO than object.__new__). This can be
reverted when FileSystem and BlockDevice classes are merged in future
and multiple inheritance is no longer necessary.

BlockDevice.filesystem_type needed implementing to verify if zfs fs
(dataset) exists on particular block device and filesystem_info added
to give extended description of what filesystem objects exist (e.g.
list of datasets) and their types. These properties for detection of
existing fs are implemented in BlockDevice subclasses.

Initialize module method now required for zfs to ensure DKMS
modules are loaded before use.

  - Implement FileSystemZfs and update BlockDeviceZfs classes
  - Ignore target.inode checks for zfs backfs
  - Implement device_matches in FileSystem{Zfs,Ldiskfs}
  - Use uuids of datasets, not zpools, in persistent storage
  - Append UpdateDevices and new UpdateManagedTargetMount steps
    to FormatTargetJob
  - Use job_scheduler_notify.notify to update mtm & target attributes
  - Lock mtm objects related to target within FormatTargetJob
  - Inherit BlockDevice{Zfs,Linux} from FileSystem{Zfs,Ldiskfs}
  - Fix FileSystem.__new__
  - Fix blockdevice.filesystem_type
  - Add blockdevice.filesystem_info
  - Add relevant Agent unit tests for BlockDevice{Linux,Zfs} and
    FileSystem{Ldiskfs,Zfs}
  - Store agent unit test data for chroma_common in example_data.py
  - Move initialize_modules from FileSystem to BlockDevice subclasses
  - Update manage_target ('Integration tests' as per SWEBOK guide)
  - Update Manager unit tests to address the fact that device_plugin
    RPCs now occur immediately after format_target RPCs
  - Add skip_calls to mock_agent_rpc.py
  - Fix behave ('Component Interface tests' as per SWEBOK) tests:
    update mock_servers with Nid object list as 'nids' dict entry

Change-Id: I258d9a43f061557729632d8fed305d870a7829ab
Signed-off-by: tanabarr <tom.nabarro@intel.com>
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/20369
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>