HYD-5062 RH7.2 doesn't support ifconfig by default
RH7.2 does not have ifconfig install by default. The solution
chosen is to make the agent work using the ip command which is
standard on all versions of linux.

This patch changes the network detection routine to use ip instead
of ifconfig. The only complication is that ip doesn't report rx/tx.
For this reason we need to also parse /proc/net/dev which does sadly
make the code less neat than we would like.

 - Change parse to parse ip output.
 - Parse stats from /proc/net/dev.
 - Create NetworkInterface classes from the two.
 - Fix a openfile issue in the zfs code.

Change-Id: I870abfa00c46d6c5dddc139fa4b17f7672e3a616
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/17180
Tested-by: Jenkins
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>