HYD-5111 Use the %{dist} tag in the Release
To support more than one major distribution release, it makes sense
to follow the upstream protocol and and include the %{dist} macro
in the RPM's Release value.

Update the appropriate makefiles and spec files to generate the full
names including distro. Add some uninstall functionality to the 
rpm-from-tarball Makefile and generic chroma-dependencies Makefile.

 - Update Makefiles to provide uninstall functionality.
 - Update the spec and Makefiles to include dist for the agent, 
   manager, diagnostics and simulator.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I8ed72ce6c005261529c4959e194738ee1b3b33a6
Reviewed-on: http://review.whamcloud.com/17290
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>