HYD-4403 Move partition number detection to the agent
Partition numbering and naming is generally supposed to
follow the convention of a base device name followed by
a number sda1, sda2 etc. However for some devices such
as LVM2 members the symbolic link to the actual device
doesn't always follow the same protocol. For example

/dev/disk/by-id/lvm-pv-uuid-rNmTx5-DSUs-lSH5-mcMq-wZxE-e9Zg-bfZfxy
could be a symlink to
/dev/vda2

The agent passes the conical name to the manager meaning
the manager in this cases cannot deduce the number of the
partition because the information is missing.

To resolve this issue the partition number detection should
be moved to the agent because the agent has more information
available than the manager. Code wise this is just a simple
refactoring.

 - Remove partition number detection regex from the manager
   in chroma_core/plugins/linux.py
 - Add parition number detection regex to the agent in
   device_plugins/linux.py
 - Updated the test in plugins/linux.py

Change-Id: I6c848fe8305367212496e4c4cb5722a7b7fcac7e
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/17652
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>