HYD-5323 Add callback registration of listeners for ServiceControl          service actions
In order to make it possible to listen to ServiceControl actions like
start and stop services. Add a callback to make this possible.

Example usage is as follows:
ServiceControl.listen('corosync', my_function)
ServiceControl.stop_listen('corosync', my_function)

This means when anything does something to a service call corosync
my_function is called.

The idea is that if something needs to happen when a service is
stopped, started, restarted, enabled etc then some code doing that
stop, start etc should not need to know that it is required. The code
doing the action should know - it is effectively an event driven
style of implementation we are after.

  - add register, unregister and notify methods to ServiceControl
    base class
  - add enum() function to common utils to generate enums and avoid
    hard coding ints
  - class methods are used to enable multiple ServiceControl
    instances to be registered to the same service and receive
    notifications when any of these instances call the relevant
    action/service
  - starting, ending and error notifications are triggered within
    ServiceControl retry action code
  - tests added to test_service_control.py
  - ServiceControlRHX notation changed to ServiceControlELX
  - single error notification given after control retries fail

Change-Id: Iedae2af64b5038433b7d74cd1fd8bd67e904d584
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/18031
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>