HYD-4717 Cluster detection for Corosync 2
For Corosync2 IML rather thean Corosync needs to detect which
nodes are in the cluster and then configure corosync appropriately.

This ticket implements the manager side that solution. The solution
uses the fact that all corosync nodes that detect the same mcastport
are a 'cluster' and so by using the mcastport as a cluster key the
manager can assemble the cluster 'manually' using chroma-agent functions.

To make development easier it also creates a Corosync2 object and
set of manage corosync2 agent commands. The duplication in the
manager is minimal and in the agent, and we can probably reduce it a
little more in future. But for development this ticket allows the
corosync2 code to develop without breaking corosync (1) code.

This patch actually hardcodes the use of Corosync and so the resulting
build will work only on el6 with a latter patch being required to
enable the el7 objects.

 - Add Corosync2 classes to the manager, these are based on a common
   ancestor as the Corosync(1) class meaning duplication is minimized.
 - Add manage_corosync2 code to the agent
 - Add manage_corosync_common to the agent so we can pull out
   suitable common code.

Change-Id: I4e31d1921ceb776d1d3a5f8f86f29f2f523766c8
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/17605
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>