HYD-5308 Move ServiceControl class to chroma-common
In order to use the ServiceControl methods in manager code as well as
agent code we need to make it common code.

ServiceControl uses the AgentShell to provide logging features for
the agent code that is calling.

In order to make this work in both manager and agent code, first
try to import AgentShell (within agent code) and if that fails
import Shell in its place (within manager code).

Note this is not an ideal solution as we are supplying knowledge
of external context into common code which should not know about the
caller environment, further work is needed to find a way to pass in
Shell and Logger objects.

  - move service control file to chroma-common
  - change import paths in agent code to reference new location
  - update and move test file to reference the new location
  - add spacing to service_control.py to improve readability

Change-Id: I1ee932eeeb87e09d73d5b896559d8b19f799b145
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/18105
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>